replace work.py with worker_handpos.py
class name WorkerHandPos()
